Implemented a new ToastProvider component
