+++ src/components/App/App.js
@@ -1,14 +1,15 @@
 import React from 'react';
 
+import ToastProvider from '../ToastProvider/ToastProvider';
 import ToastPlayground from '../ToastPlayground';
 import Footer from '../Footer';
 
 function App() {
   return (
-    <>
+    <ToastProvider>
       <ToastPlayground />
       <Footer />
-    </>
+    </ToastProvider>
   );
 }
