+++ src/components/ToastPlayground/ToastPlayground.js
@@ -4,6 +4,7 @@ import Button from '../Button';
 
 import styles from './ToastPlayground.module.css';
 
+import { ToastContext } from '../ToastProvider/ToastProvider';
 import ToastShelf from '../ToastShelf/ToastShelf';
 
 const VARIANT_OPTIONS = ['notice', 'warning', 'success', 'error'];
@@ -11,31 +12,15 @@ const VARIANT_OPTIONS = ['notice', 'warning', 'success', 'error'];
 function ToastPlayground() {
   const [message, setMessage] = React.useState();
   const [variant, setVariant] = React.useState('notice');
-  const [toasts, setToasts] = React.useState([]);
 
-  function handleAddToast(message, variant) {
-    const newToast = {
-      id: Math.random(),
-      message,
-      variant,
-    };
-    const nextToast = [...toasts, newToast];
-    setToasts(nextToast);
-  }
+  const { toasts } = React.useContext(ToastContext);
+  const { handleAddToast } = React.useContext(ToastContext);
 
   function handleSubmit(event) {
     event.preventDefault();
     handleAddToast(message, variant);
     setMessage('');
     setVariant('notice');
-  }
-
-  function handleCloseById(id) {
-    const nextToasts = toasts.filter((toast) => {
-      return toast.id !== id;
-    });
-
-    setToasts(nextToasts);
   }
 
   return (
@@ -45,9 +30,7 @@ function ToastPlayground() {
           <img alt="Cute toast mascot" src="/toast.png" />
           <h1>Toast Playground</h1>
         </header>
-        {toasts.length > 0 && (
-          <ToastShelf toasts={toasts} closeToastById={handleCloseById} />
-        )}
+        {toasts.length > 0 && <ToastShelf toasts={toasts} />}
         <div className={styles.controlsWrapper}>
           <div className={styles.row}>
             <label
