+++ src/components/ToastProvider/index.js
@@ -1,0 +1,2 @@
+export * from './ToastProvider';
+export { default } from './ToastProvider';
