+++ src/components/ToastShelf/ToastShelf.js
@@ -1,23 +1,22 @@
 import React from 'react';
 
+import { ToastContext } from '../ToastProvider/ToastProvider';
 import Toast from '../Toast';
 import styles from './ToastShelf.module.css';
 
-function ToastShelf({ toasts, closeToastById }) {
-  if (toasts.length === 0) return null;
+function ToastShelf() {
+  const { toasts } = React.useContext(ToastContext);
 
-  function onToastClose(id) {
-    closeToastById(id);
-  }
+  if (toasts.length === 0) return null;
 
   return (
     <ol className={styles.wrapper}>
       {toasts.map((toast) => (
         <li key={toast.id} className={styles.toastWrapper}>
           <Toast
+            id={toast.id}
             message={toast.message}
             variant={toast.variant}
-            onClose={() => onToastClose(toast.id)}
           />
         </li>
       ))}
